BitFit: Simple Parameter-efficient Fine-tuning for Transformer-based Masked Language-models
We show that with small-to-medium training
data, fine-tuning only the bias terms (or a subset of the bias terms) of pre-trained BERT
models is competitive with (and sometimes
better than) fine-tuning the entire model. For
larger data, bias-only fine-tuning is competitive with other sparse fine-tuning methods.
Besides their practical utility, these findings
are relevant for the question of understanding the commonly-used process of finetuning:
they support the hypothesis that finetuning is
mainly about exposing knowledge induced by
language-modeling training, rather than learning new task-specific linguistic knowledge.