Discerning Generic Event Boundaries in Long-Form Wild Videos
Detecting generic, taxonomy-free event boundaries in
videos represents a major stride forward towards holistic
video understanding. In this paper we present a technique for
generic event boundary detection based on a two stream inflated 3D convolutions architecture, which can learn spatiotemporal features from videos. Our work is inspired from the
Generic Event Boundary Detection Challenge (part of CVPR
2021 Long Form Video Understanding- LOVEU Workshop).
Throughout the paper we provide an in-depth analysis of
the experiments performed along with an interpretation of
the results obtained. 

The code for this work can be found at
https://github.com/rayush7/GEBD