Medical Image Analysis on Left Atrial LGE MRI for Atrial Fibrillation Studies: A Review
Late gadolinium enhancement magnetic resonance imaging (LGE MRI) is commonly
used to visualize and quantify left atrial (LA) scars. The position and extent of scars provide important information of the pathophysiology and progression of atrial fibrillation
(AF). Hence, LA scar segmentation and quantification from LGE MRI can be useful in
computer-assisted diagnosis and treatment stratification of AF patients. Since manual
delineation can be time-consuming and subject to intra- and inter-expert variability, automating this computing is highly desired, which nevertheless is still challenging and
under-researched.
This paper aims to provide a systematic review on computing methods for LA cavity, wall, scar and ablation gap segmentation and quantification from LGE MRI, and
the related literature for AF studies. Specifically, we first summarize AF-related imaging techniques, particularly LGE MRI. Then, we review the methodologies of the four
computing tasks in detail, and summarize the validation strategies applied in each task.
Finally, the possible future developments are outlined, with a brief survey on the potential clinical applications of the aforementioned methods. The review shows that the
research into this topic is still in early stages. Although several methods have been proposed, especially for LA segmentation, there is still large scope for further algorithmic
developments due to performance issues related to the high variability of enhancement
appearance and differences in image acquisition.